+++ dependencies/Gaff/include/Gaff_IReflectionDefinitions.h
@@ -42,7 +42,7 @@ public:
 	virtual unsigned int getValueGeneric(const char* name) const = 0;
 
 	virtual Pair<AString<Allocator>, unsigned int> getEntryGeneric(unsigned int index) const = 0;
-	virtual unsigned int getNumEntries(void) const = 0;
+	virtual size_t getNumEntries(void) const = 0;
 
 	virtual const char* getEnumName(void) const = 0;
 };
