+++ dependencies/Gleam/include/Gleam_ProxyAllocator.h
@@ -35,7 +35,7 @@ NS_GLEAM
 class ProxyAllocator : public Gaff::IAllocator
 {
 public:
-	void* alloc(unsigned int size_bytes)
+	void* alloc(size_t size_bytes)
 	{
 		return GetAllocator()->alloc(size_bytes);
 	}
